mtd: spi-nor: fix silent truncation in spi_nor_read_raw() (bsc#1175416).
suse-commit: 6ae42b4acb08a156d36a88d166984a3e360f46dd

+++ drivers/mtd/spi-nor/spi-nor.c
@@ -1794,7 +1794,7 @@ static int spi_nor_read_sfdp(struct spi_nor *nor, u32 addr,
 			     size_t len, void *buf)
 {
 	u8 addr_width, read_opcode, read_dummy;
-	int ret;
+	ssize_t ret;
 
 	read_opcode = nor->read_opcode;
 	addr_width = nor->addr_width;
